backend/requirements.txt
# Flask and core dependencies
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-Mail==0.9.1

# Database
PyMySQL==1.1.0
SQLAlchemy==2.0.23
mysql-connector-python==8.2.0
cryptography==41.0.7

# Environment and configuration
python-dotenv==1.0.0
python-decouple==3.8

# Security and validation
bcrypt==4.1.2
Werkzeug==3.0.1
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0

# Utilities
requests==2.31.0
Pillow==10.1.0
python-dateutil==2.8.2

# Development dependencies
pytest==7.4.3
pytest-flask==1.3.0
black==23.11.0
flake8==6.1.0

# Production server
gunicorn==21.2.0
